+++ service/job/internal/src/main/java/org/eclipse/kapua/service/job/execution/internal/JobExecutionModule.java
@@ -1,0 +1,25 @@
+/*******************************************************************************
+ * Copyright (c) 2021 Eurotech and/or its affiliates and others
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     Eurotech - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.kapua.service.job.execution.internal;
+
+import org.eclipse.kapua.commons.core.AbstractKapuaModule;
+import org.eclipse.kapua.service.job.execution.JobExecutionFactory;
+import org.eclipse.kapua.service.job.execution.JobExecutionService;
+
+public class JobExecutionModule extends AbstractKapuaModule {
+    @Override
+    protected void configureModule() {
+        bind(JobExecutionFactory.class).to(JobExecutionFactoryImpl.class);
+        bind(JobExecutionService.class).to(JobExecutionServiceImpl.class);
+    }
+}
