+++ service/job/internal/src/main/java/org/eclipse/kapua/service/job/internal/JobCreatorImpl.java
@@ -18,7 +18,6 @@ import org.eclipse.kapua.service.job.Job;
 import org.eclipse.kapua.service.job.JobCreator;
 import org.eclipse.kapua.service.job.step.JobStep;
 
-import javax.inject.Singleton;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -27,7 +26,6 @@ import java.util.List;
  *
  * @since 1.0.0
  */
-@Singleton
 public class JobCreatorImpl extends AbstractKapuaNamedEntityCreator<Job> implements JobCreator {
 
     private static final long serialVersionUID = 3119071638220738358L;
