+++ service/job/internal/src/main/java/org/eclipse/kapua/service/job/step/internal/JobStepCreatorImpl.java
@@ -18,7 +18,6 @@ import org.eclipse.kapua.service.job.step.JobStep;
 import org.eclipse.kapua.service.job.step.JobStepCreator;
 import org.eclipse.kapua.service.job.step.definition.JobStepProperty;
 
-import javax.inject.Singleton;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -27,7 +26,6 @@ import java.util.List;
  *
  * @since 1.0.0
  */
-@Singleton
 public class JobStepCreatorImpl extends AbstractKapuaNamedEntityCreator<JobStep> implements JobStepCreator {
 
     private static final long serialVersionUID = 3119071638220738358L;
