+++ service/job/internal/src/main/java/org/eclipse/kapua/service/job/targets/internal/JobTargetCreatorImpl.java
@@ -17,9 +17,6 @@ import org.eclipse.kapua.model.id.KapuaId;
 import org.eclipse.kapua.service.job.targets.JobTarget;
 import org.eclipse.kapua.service.job.targets.JobTargetCreator;
 
-import javax.inject.Singleton;
-
-@Singleton
 public class JobTargetCreatorImpl extends AbstractKapuaUpdatableEntityCreator<JobTarget> implements JobTargetCreator {
 
     private static final long serialVersionUID = 3119071638220738358L;
